+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/entity/oddion/OddionRenderer.java
@@ -1,8 +1,6 @@
 package net.potionstudios.biomeswevegone.world.entity.oddion;
 
-import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.model.geom.ModelLayerLocation;
-import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import org.jetbrains.annotations.NotNull;
@@ -23,8 +21,7 @@ public class OddionRenderer<T extends Oddion> extends GeoEntityRenderer<T> {
     }
 
     @Override
-    public void render(@NotNull T entity, float entityYaw, float partialTicks, @NotNull PoseStack stack, @NotNull MultiBufferSource bufferIn, int packedLightIn) {
-        shadowRadius = 0.4f;
-        super.render(entity, entityYaw, partialTicks, stack, bufferIn, packedLightIn);
+    protected float getShadowRadius(@NotNull T entity) {
+        return 0.4f;
     }
 }